requirements.txt
appdirs==1.4.4
asgiref==3.4.1
bcrypt==3.2.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
DateTime==4.3
distlib==0.3.1
Django==3.2.7
django-cors-headers==3.7.0
django-storages==1.11.1
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
h11==0.12.0
httplib2==0.19.1
idna==2.10
jmespath==0.10.0
mysqlclient==2.0.3
oauthlib==3.1.0
packaging==20.9
phonenumbers==8.12.23
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.2
starlette==0.13.6
text-unidecode==1.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.5
uvicorn==0.13.4
virtualenv==20.4.3
zope.interface==5.4.0
whitenoise==5.3.0